# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RegistrationInput {
  /** @minLength 2 */
  name: string;
  /** @pattern ^[^@\s]+@[^@\s]+\.[^@\s]+$ */
  email: string;
  /** @minLength 2 */
  school: string;
  grade: string;
  /** @nullable */
  teamName?: string | null;
  /**
     * @minimum 1
     * @maximum 4
     */
  teamSize: number;
  experience: string;
  /** @nullable */
  projectIdea?: string | null;
  /** @nullable */
  dietaryNeeds?: string | null;
}

export type Registration = RegistrationInput & {
  id: number;
  createdAt: string;
};

export interface Announcement {
  id: number;
  title: string;
  body: string;
  label: string;
  publishedAt: string;
  isPinned: boolean;
}

export interface AnnouncementInput {
  /** @minLength 2 */
  title: string;
  /** @minLength 2 */
  body: string;
  /** @minLength 2 */
  label: string;
  isPinned: boolean;
}

export interface AnnouncementUpdate {
  id: number;
  /** @minLength 2 */
  title: string;
  /** @minLength 2 */
  body: string;
  /** @minLength 2 */
  label: string;
  isPinned: boolean;
}

export type ScheduleItemKind = typeof ScheduleItemKind[keyof typeof ScheduleItemKind];


export const ScheduleItemKind = {
  event: 'event',
  workshop: 'workshop',
  break: 'break',
  judging: 'judging',
} as const;

export interface ScheduleItem {
  id: number;
  startTime: string;
  endTime: string;
  title: string;
  description: string;
  kind: ScheduleItemKind;
  location: string;
}

export interface DashboardSummary {
  registrationCount: number;
  teamCount: number;
  announcementCount: number;
  nextEvent: ScheduleItem;
}

export type AdminKeyParameter = string;