# lib/api-zod/src/generated/types/announcement.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Announcement {
  id: number;
  title: string;
  body: string;
  label: string;
  publishedAt: Date;
  isPinned: boolean;
}