# lib/api-zod/src/generated/types/announcementUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AnnouncementUpdate {
  id: number;
  /** @minLength 2 */
  title: string;
  /** @minLength 2 */
  body: string;
  /** @minLength 2 */
  label: string;
  isPinned: boolean;
}