# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScheduleItem } from './scheduleItem';

export interface DashboardSummary {
  registrationCount: number;
  teamCount: number;
  announcementCount: number;
  nextEvent: ScheduleItem;
}