# lib/api-zod/src/generated/types/registration.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RegistrationInput } from './registrationInput';

export type Registration = RegistrationInput & {
  id: number;
  createdAt: Date;
};