# lib/api-zod/src/generated/types/registrationInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RegistrationInput {
  /** @minLength 2 */
  name: string;
  /** @pattern ^[^@\s]+@[^@\s]+\.[^@\s]+$ */
  email: string;
  /** @minLength 2 */
  school: string;
  grade: string;
  /** @nullable */
  teamName?: string | null;
  /**
     * @minimum 1
     * @maximum 4
     */
  teamSize: number;
  experience: string;
  /** @nullable */
  projectIdea?: string | null;
  /** @nullable */
  dietaryNeeds?: string | null;
}