# lib/api-zod/src/generated/types/scheduleItem.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScheduleItemKind } from './scheduleItemKind';

export interface ScheduleItem {
  id: number;
  startTime: string;
  endTime: string;
  title: string;
  description: string;
  kind: ScheduleItemKind;
  location: string;
}